app/requirements.txt
psycopg2
pymongo
PyYAML